+++ user-service/src/main/java/com/poscoict/mes/user/admin/jpa/AdminProductEntity.java
@@ -22,8 +22,8 @@ public class AdminProductEntity {
 	
 	private String productId;
 	private Double thickness;
-	private Integer width;
-	private Integer length;
+	private Double width;
+	private Double length;
 	private String productStatus;
 	private String productUpdate;
